made changes in #7
