--- frontend/sign-up-page/README.md
@@ -1,19 +1,0 @@
-# Issue # 7 : Setup Sign Up Page
-
-This project was bootstrapped with [Create React App](https://github.com/facebook/create-react-app).
-
-## how to run
-1. clone the repo
-2. go to the cloned repo folder
-3. npm install
-4. npm start
-5. open http://localhost:3000/ in your browser
-
-### Screenshots of the frontend :
-
-![img-1](image.png)
-
-<br>
-<br>
-
-![img-2](image-1.png)
