--- frontend/sign-up-page/public/index.html
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <meta name="description" content="Create Account Page with Profile Picture Upload">
-  <title>Create Account</title>
-  <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Comic+Sans+MS:wght@400;700&display=swap">
-</head>
-<body>
-  <div id="root"></div>
-</body>
-</html>
