--- frontend/sign-up-page/src/App.css
@@ -1,4 +1,0 @@
-/* src/App.css */
-.app {
-  font-family: 'Comic Sans MS', cursive, sans-serif;
-}
