--- frontend/sign-up-page/src/components/Background.css
@@ -1,31 +1,0 @@
-/* src/components/Background.css */
-.background {
-    background-color: #254c44;
-    min-height: 1200px;
-    padding: 0rem;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-  }
-  
-  
-  .background::before {
-    content: '';
-    position:fixed;
-    margin-top: 80px;
-    margin-left:0px;
-    top: 10px;
-    left: 2px;
-    bottom: 0px;
-    width: 36%; /* Adjust the width of the image as needed */
-    background-image: url('./bg-image.png'); /* URL of the background image */
-    background-size: cover; /* Cover the entire area */
-    background-position: left ;
-    z-index: +1; 
-  }
-  
-  @media (max-width: 768px) {
-    .background::before {
-      width: 100%; 
-    }
-  }
