--- frontend/sign-up-page/src/components/Background.jsx
@@ -1,9 +1,0 @@
-// src/components/Background.jsx
-import React from 'react';
-import './Background.css';
-
-const Background = ({ children }) => {
-  return <div className="background">{children}</div>;
-};
-
-export default Background;
