--- frontend/sign-up-page/src/components/CreateAccountForm.css
@@ -1,119 +1,0 @@
-/* src/components/CreateAccountForm.css */
-.create-account-form {
-    
-    color: #b07156;
-    padding-top: 4rem;
-    border-radius: 8px;
-    text-align: center;
-    max-width: 600px;
-    width: 100%;
-    margin-left: 45%; /* Center the form horizontally */
-  }
-  
-  .create-account-form h1 {
-    font-size: 3.7rem;
-    font-weight: 700px;
-    margin-bottom: 2rem;
-    text-shadow: -3px 0px 0px #000, 1px -1px 0 #000, -1px 1px 0 #000, 1px 1px 0 #000;
-    
-  }
-  
-  .create-account-form form {
-    display: grid;
-    padding-top: 2rem;
-    grid-template-columns: 1fr 1fr;
-    font-size: 1.2rem;
-    gap: 3rem;
-    color: #22e8b9;
-    text-shadow: -1px -1px 0 #000, 1px -1px 0 #000, -1px 1px 0 #000, 1px 1px 0 #000;
-    text-align: left;
-    font-family: 'Comic Sans MS', cursive, sans-serif;
-  }
-  
-  .create-account-form .form-group {
-    display: flex;
-    flex-direction: column;
-    gap: 1.3rem;
-  }
-  
-  .create-account-form .form-group label {
-    margin-bottom: 0.5rem;
-    font-weight: bold;
-  }
-  
-  .create-account-form .form-group input,
-  .create-account-form .form-group select {
-    padding: 0.6rem;
-    border: none;
-    border-radius: 10px;
-    font-family: 'Comic Sans MS', cursive, sans-serif;
-    background-color: #87BBA2;
-    border: 3px solid black;
-  }
-  
-  .create-account-form .form-group:nth-child(6) {
-    grid-column: 1 / -1;
-  }
-  
-  .create-account-form .form-group:last-child {
-    grid-column: 1 / -1;
-  }
-  
-  .create-account-form button {
-    font-family: 'Comic Sans MS', cursive, sans-serif;
-    font-size: medium;
-    grid-column: 1 / -1;
-    border: 3px solid black;
-    border-radius: 6px;
-    background-color: #ffcc00;
-    color: #000;
-    font-weight: bold;
-    cursor: pointer;
-    height: 40px;
-    margin-left: 260px;
-    width: fit-content;
-    padding:fit-content;
-    
-  }
-  
-  .create-account-form .google-signup button {
-    position: relative; /* Ensure relative positioning for the button */
-    padding: 0.75rem 2rem 0.75rem 0.75rem; /* Adjust padding as needed */
-    border: none;
-    border-radius: 10px;
-    background-color: #ffcc00;
-    color: #000;
-    font-weight: bold;
-    cursor: pointer;
-    border: 3px solid black;
-    margin-top: 10px;
-    margin-left: -10px;
-  }
-  
-  .create-account-form .google-signup button::after {
-    content: ''; /* Inserting a pseudo-element for the image */
-    position: absolute;
-    width: 24px; /* Adjust width and height as needed */
-    height: 20px;
-    background-image: url('./google-logo.png'); 
-    background-size: contain; 
-    background-repeat: no-repeat;
-    background-position: center;
-    right: 5px; 
-    top: 50%;
-    transform: translateY(-30%);
-    
-  }
-  
-  .create-account-form p {
-    margin: 1rem 0;
-    grid-column: 1 / -1; /* Make the paragraph span both columns */
-  }
-  
-  .error {
-    color: red;
-    font-size: 0.875rem;
-    margin-top: -0.75rem;
-  }
-  
-
