--- frontend/sign-up-page/src/components/Navbar.css
@@ -1,43 +1,0 @@
-/* src/components/Navbar.css */
-
-.navbar {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding: 1rem 2rem;
-    background-color:#20AD9A;    
-    color: #fff;
-    width: 100%;
-    height: 47px;
-    position: fixed;
-    top: 0;
-    left: 0;
-    z-index:1;
-  }
-  
-  .navbar-logo {
-    font-size: 2rem;
-    font-weight: bold;
-  }
-  
-  .navbar-links {
-    display: flex;
-    gap: 0.5 rem;
-  }
-  
-  .navbar-links a {
-    margin-left: 50px;
-    color: #fff;
-    text-decoration: none;
-    font-size: 1.3rem;
-    transition: color 0.3s ease;
-  }
-  
-  .navbar-links a:hover {
-    color: #ffcc00;
-  }
-  .navbar-user-icon img {
-    margin-right:50px;
-    width:40px;
-  }
-
