--- frontend/sign-up-page/src/components/ProfilePictureUpload.css
@@ -1,17 +1,0 @@
-/* src/components/ProfilePictureUpload.css */
-.profile-picture-upload {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    margin: 1rem 0;
-  }
-  
-  .profile-picture-upload input {
-    margin-bottom: 1rem;
-  }
-  
-  .profile-picture-upload img {
-    max-width: 150px;
-    border-radius: 50%;
-  }
-
