--- frontend/sign-up-page/src/index.css
@@ -1,16 +1,0 @@
-/* src/index.css */
-body {
-  margin: 0;
-  font-family: 'Comic Sans MS', cursive, sans-serif;
-  background-color: #254c44;
-  color: #fff;
-}
-
-#root {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  min-height: 100vh;
-  padding: 2rem;
-}
