+++ frontend/src/App.js
@@ -3,6 +3,7 @@ import { Routes, Route } from 'react-router-dom';
 import HomePage from './components/Home/HomePage';
 import Generate from './components/Generate/generate'; // Ensure this path is correct
 import Navbar from './components/Nav/Navbar';
+import CreateAccountForm from './components/sign_up/CreateAccountForm';
 import './App.css';
 
 function App() {
@@ -12,7 +13,8 @@ function App() {
       <Routes>
         <Route path="/" element={<HomePage />} />
         <Route path="/generate" element={<Generate />} />
-      </Routes>
+        <Route path="/CreateAccountForm" element={<CreateAccountForm/>} />
+        </Routes>
     </div>
   );
 }
